requirements.txt
Django>=4.2.0,<5.0.0
django-cors-headers>=4.3.0
django-filter>=23.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
jsonfield>=3.1.0
PyJWT>=2.8.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0
whitenoise>=6.6.0 